fixed the build error
