+++ src/app/Contact Section/page.tsx
@@ -1,14 +1,7 @@
 "use client";
 
-// import Image from "next/image";
-// import Link from "next/link";
+import { ContactSection } from "@/components/sections/ContactSection";
 
-import { ContactSectionBackground } from "./contact-section-background";
-
-export const ContactSection = () => {
-	return (
-		<div id="ContactSection" className=" overflow-hidden">
-			<ContactSectionBackground />
-		</div>
-	);
-};
+export default function ContactPage() {
+	return <ContactSection />;
+}
