+++ src/app/Services Section/page.tsx
@@ -1,14 +1,6 @@
 "use client";
-import { ServiceCards } from "@/components/ServiceCards";
-import { AuroraBackground } from "@/components/ui/aurora-background";
+import { ServicesSection } from "@/components/sections/ServicesSection";
 
-export function ServicesSection() {
-	return (
-		<AuroraBackground id="ServicesSection" className="w-full pt-28 pb-10 h-fit">
-			<h1 className="text-center text-5xl font-bold text-white">
-				Our Services
-			</h1>
-			<ServiceCards />
-		</AuroraBackground>
-	);
+export default function ServicesPage() {
+	return <ServicesSection />;
 }
