+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { Navbar } from "@/components/navbar";
 
 const geistSans = Geist({
 	variable: "--font-geist-sans",
@@ -15,30 +14,7 @@ const geistMono = Geist_Mono({
 
 export const metadata: Metadata = {
 	title: "VITrendz",
-	description:
-		"VITrendz, Your go-to resource hub for VIT college students featuring free academic tools like FFCS planner, CGPA calculator, and more to simplify campus life.",
-	keywords:
-		"VIT, VIT Vellore, FFCS planner, CGPA calculator, Resources, Hostels, Roommate, Vtrendz, VITrendz, FFCS, VIT FFCS, VIT CGPA, VIT Resources, VIT Hostels, VIT Roommate, Feedback, Student feedback",
-	authors: [{ name: "VTech team" }],
-	creator: "VTech Team",
-	publisher: "VITrendz",
-	formatDetection: {
-		email: false,
-		address: false,
-		telephone: false,
-	},
-	metadataBase: new URL("https://vitrendz.in"),
-	alternates: {
-		canonical: "/",
-	},
-	viewport: {
-		width: "device-width",
-		initialScale: 1,
-	},
-	robots: {
-		index: true,
-		follow: true,
-	},
+	description: "Our description",
 };
 
 export default function RootLayout({
@@ -51,7 +27,6 @@ export default function RootLayout({
 			<body
 				className={`${geistSans.variable} ${geistMono.variable} antialiased`}
 			>
-				<Navbar />
 				{children}
 			</body>
 		</html>
