+++ src/app/page.tsx
@@ -1,17 +1,21 @@
-import { HeroSection } from "@/app/Hero section/page";
-import { AboutSection } from "./About Section/page";
-import { ServicesSection } from "./Services Section/page";
-import { CommunitySection } from "./Community Section/page";
-import { ContactSection } from "./Contact Section/page";
+import { HeroSection } from "@/components/sections/HeroSection";
+import { AboutSection } from "@/components/sections/AboutSection";
+import { Navbar } from "../components/navbar";
+import { ServicesSection } from "@/components/sections/ServicesSection";
+import { CommunitySection } from "@/components/sections/CommunitySection";
+import { ContactSection } from "@/components/sections/ContactSection";
 
 export default function Home() {
 	return (
+		<>
 		<div className="bg-black text-white overflow-hidden">
+			<Navbar />
 			<HeroSection />
 			<AboutSection />
 			<ServicesSection />
 			<CommunitySection />
 			<ContactSection />
 		</div>
+		</>
 	);
 }
