+++ src/components/sections/CommunitySection.tsx
@@ -1,0 +1,12 @@
+import { CommunityTabs } from "@/components/CommunityTabs";
+
+export function CommunitySection() {
+	return (
+		<>
+				<h1 className="text-center text-5xl text-white bg-black font-bold -mb-20 mt-20">
+					Our Community
+				</h1>
+				<CommunityTabs />
+		</>
+	);
+}
