+++ src/components/sections/ContactSection.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { ContactSectionBackground } from "@/app/Contact Section/contact-section-background";
+
+export function ContactSection() {
+	return (
+		<div id="ContactSection" className=" overflow-hidden">
+			<ContactSectionBackground />
+		</div>
+	);
+}
