+++ src/components/sections/HeroSection.tsx
@@ -1,0 +1,9 @@
+import { HeroSectionBackground } from "@/app/Hero section/hero-section";
+
+export function HeroSection() {
+	return (
+		<div className="overflow-hidden">
+			<HeroSectionBackground />
+		</div>
+	);
+}
