+++ src/components/ui/carousel.tsx
@@ -1,5 +1,6 @@
 "use client";
 import { IconArrowNarrowRight } from "@tabler/icons-react";
+import Image from "next/image";
 import Link from "next/link";
 // import Image from "next/image";
 import { useState, useRef, useId, useEffect } from "react";
@@ -23,7 +24,7 @@ const Slide = ({ slide, index, current, handleSlideClick }: SlideProps) => {
 
 	const xRef = useRef(0);
 	const yRef = useRef(0);
-	const frameRef = useRef<number>();
+	const frameRef = useRef<number | undefined>(undefined);
 
 	useEffect(() => {
 		const animate = () => {
@@ -93,7 +94,7 @@ const Slide = ({ slide, index, current, handleSlideClick }: SlideProps) => {
 								: "none",
 					}}
 				>
-					<img
+					<Image
 						className="absolute inset-0 w-[120%] h-[120%] object-cover opacity-100 transition-opacity duration-600 ease-in-out"
 						style={{
 							opacity: current === index ? 1 : 0.5,
