+++ src/components/ui/tabs.tsx
@@ -7,7 +7,7 @@ import { cn } from "@/lib/utils";
 type Tab = {
 	title: string;
 	value: string;
-	content?: string | React.ReactNode | any;
+	content?: React.ReactNode;
 };
 
 export const Tabs = ({
