+++ src/components/ui/text-generate-effect.tsx
@@ -15,7 +15,7 @@ export const TextGenerateEffect = ({
 	duration?: number;
 }) => {
 	const [scope, animate] = useAnimate();
-	let wordsArray = words.split(" ");
+	const wordsArray = words.split(" ");
 	useEffect(() => {
 		animate(
 			"span",
